[Darwin] Amend handling of ptyopen names
On some versions of Darwin, it seems that ttyname() might return null, this causes
expect and thence runtest to segv, sometimes in a hard-to-spot place in large test
suites.

It seems that, in those cases, there's a useful name returned from openpty.
So - try the name from openpty, if that's not present, try ttyname(slave).
If that's not present exit with an error.