[Darwin] Amend handling of ptyopen() slave name.
The man page says:
int
     openpty(int *amaster, int *aslave, char *name, struct termios *termp, struct winsize *winp);

    " If name is non-null, the filename of the slave is returned in name. "

For some reason, in the original code, that was being copied into the master name.
The slave FID was then being queried for its ttyname (which sometimes fails; possibly an OS bug,).
Anyway, the upshot is that the best thing to do is to point the openpty() call to the slave name and if that fails to be set, make a last ditch attempt to collect it using ttyname() on the slave FID.  If that fails we print a message and bail.